End-to-end release test (#406)
We currently have an end-to-end test in place to test the latest code in the repo. This PR adds a second test that tests our latest release. It scrapes the version tags from the public website (testing that the website is live in the process), and then uses those values to launch the rest of our e2e tests.

This one isn't tied to any code changes, so I set it to run daily and manually when requested