update SLO for rating service
exclude loading requests from availability and latency SLOs of the rating
service. due to lack of support for mixing 'AND' and 'OR' operators in
metric labels filter the whole part that excludes responses with HTTP
status in between 400 and 500 is commented out until there will be a
solution.
latency SLO is brought in sync with latency SLOs of other services (500ms).